% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.questionnaire.R
\name{adjust_total_score_bounds}
\alias{adjust_total_score_bounds}
\title{Adjusts provided total min/max to be within mathematically possible bounds for a questionnaire.}
\usage{
adjust_total_score_bounds(total_min, total_max, n_items, likert, q_name = NULL)
}
\arguments{
\item{total_min}{User-specified minimum total score.}

\item{total_max}{User-specified maximum total score.}

\item{n_items}{Number of items in the questionnaire.}

\item{likert}{Vector of possible Likert scale values.}

\item{q_name}{(Optional) Name of the questionnaire for warnings.}
}
\value{
A list containing the adjusted total_min and total_max.
}
\description{
Adjusts provided total min/max to be within mathematically possible bounds for a questionnaire.
}
