% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-antilogit.R
\name{antilogit}
\alias{antilogit}
\title{Anti-logit (Logistic) Transformation}
\usage{
antilogit(x)
}
\arguments{
\item{x}{Numeric value or vector of real numbers}
}
\value{
Probability value(s) between 0 and 1
}
\description{
Calculates the inverse logit (logistic) transformation of a real number.
This function transforms any real number into a probability between 0 and 1.
}
\examples{
antilogit(0)    # Returns 0.5
antilogit(1.099) # Returns ~0.75

}
