% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap.pred.r
\name{ap.pred}
\alias{ap.pred}
\title{Predict "approaching parameter" (as in Brenner and Smeets' model (1998))}
\usage{
ap.pred(MGA, r)
}
\arguments{
\item{MGA}{maximum grip aperture}

\item{r}{radius of the target object}
}
\description{
Predict "approaching parameter" (as in Brenner and Smeets' model (1998))
}
\details{
By definition, MGA must be >= r*2
}
\examples{
libraries()
(ap <- ap.pred(MGA=100.25, r=40))
}
