% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{cc.derive}
\alias{cc.derive}
\title{smooth then calculate derivatives of a given variable with respect to a specific predictor. Inputs:}
\usage{
cc.derive(x, y, l = 0.2, d)
}
\arguments{
\item{x:}{predictor}

\item{y:}{variable to be smoothed}

\item{l:}{lambda - smoothing parameter (default to .2)}

\item{d:}{order of derivative}
}
\description{
smooth then calculate derivatives of a given variable with respect to a specific predictor. Inputs:
}
