% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{cc.max}
\alias{cc.max}
\title{Maximum with NA Handling}
\usage{
cc.max(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Maximum value in the vector, ignoring NA values
}
\description{
Wrapper for max() with na.rm=TRUE by default. Useful for ddply and graphs.
}
\examples{
cc.max(c(1,2,NA,4)) # Returns 4

}
