% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{cc.mean}
\alias{cc.mean}
\title{Mean with NA Handling}
\usage{
cc.mean(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Mean value of the vector, ignoring NA values
}
\description{
Wrapper for mean() with na.rm=TRUE by default
}
\examples{
cc.mean(c(1,2,NA,4)) # Returns 2.333...

}
