% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{cc.min}
\alias{cc.min}
\title{Minimum with NA Handling}
\usage{
cc.min(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Minimum value in the vector, ignoring NA values
}
\description{
Wrapper for min() with na.rm=TRUE by default. Useful for ddply and graphs.
}
\examples{
cc.min(c(1,2,NA,4)) # Returns 1

}
