% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{cc.smooth}
\alias{cc.smooth}
\title{smooth data points.}
\usage{
cc.smooth(x, y, l = 0.2)
}
\arguments{
\item{x:}{predictor}

\item{y:}{variable to be smoothed}

\item{l:}{lambda - smoothing parameter (default to .2)}
}
\description{
smooth data points.
}
