% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.exp.r
\name{des.exp}
\alias{des.exp}
\title{Tool for building experiment designs}
\usage{
des.exp(dv.n, iv.lev, formula = "full", is.iv.factor)
}
\description{
Returns the regressors matrix given a certain set of independent variables. See ?sim.mvlmod for details.
}
