% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grip.aperture.r
\name{grip.aperture}
\alias{grip.aperture}
\title{Generate grip aperture profiles accordind to Brenner and Smeets' model (1998)}
\usage{
grip.aperture(ap, MT, r)
}
\arguments{
\item{ap}{approaching parameter, as in Brenner & Smeets (1998)}

\item{MT}{movement time (in millisec)}

\item{r}{radius of the target object}
}
\description{
Generate grip aperture profiles accordind to Brenner and Smeets' model (1998)
}
\examples{
libraries()
mov_time <- 1500
GA <- grip.aperture(ap=1500, MT=mov_time, r=40)
time <- 1:mov_time
qplot(time, GA)
}
