% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{lengthunique}
\alias{lengthunique}
\title{Count Unique Elements}
\usage{
lengthunique(x)
}
\arguments{
\item{x}{A vector of any type}
}
\value{
Integer representing the count of unique elements
}
\description{
Returns the number of unique elements in a vector
}
\examples{
lengthunique(c(1,1,2,3,3,4)) # Returns 4
lengthunique(c("a","a","b","c")) # Returns 3

}
