% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{lmax.locate}
\alias{lmax.locate}
\title{Locate Local Maxima}
\usage{
lmax.locate(x, y)
}
\arguments{
\item{x}{Numeric vector of x coordinates}

\item{y}{Numeric vector of y values}
}
\value{
Logical vector indicating which x values correspond to local maxima
}
\description{
Identifies local maxima in a sequence of y values corresponding to x coordinates.
Used in kin.extract function.
}
\examples{
x <- 1:5
y <- c(1,2,3,2,1)
lmax.locate(x, y) # Returns c(FALSE,FALSE,TRUE,FALSE,FALSE)

}
