% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-antilogit.R
\name{logit}
\alias{logit}
\title{Logit Transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{Numeric value or vector of probabilities between 0 and 1}
}
\value{
The logit-transformed value(s)
}
\description{
Calculates the logit transformation of a probability (p), defined as log(p/(1-p)).
The logit function is the inverse of the logistic function.
}
\examples{
logit(0.5)  # Returns 0
logit(0.75) # Returns ~1.099

}
