% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGA.calc.r
\name{MGA.calc}
\alias{MGA.calc}
\title{Calculate maximum grip aperture accordind to Brenner and Smeets' model (1998)}
\usage{
MGA.calc(ap, r)
}
\arguments{
\item{ap}{approaching parameter, as in Brenner & Smeets (1998)}

\item{r}{radius of the target object}
}
\description{
Calculate maximum grip aperture accordind to Brenner and Smeets' model (1998)
}
\examples{
# single calculation
(MGA <- MGA.calc(ap=1500, r=40))

# it is possible to explore the range of values of ap 
# (given a certain radius) as a function of MGA
libraries()
app.par <- 100:1000
rad <- 20
MGA <- MGA.calc(ap=app.par, r=rad)
qplot(MGA, app.par)
}
