% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA_utils.R
\name{multi}
\alias{multi}
\title{Extract Multivariate Results from ANOVA Object}
\usage{
multi(ANOVA.object)
}
\arguments{
\item{ANOVA.object}{An object of class "ANOVA" containing analysis results}
}
\value{
A data frame containing multivariate test results including Pillai's Trace,
Wilks' Lambda, Hotelling's Trace, and Roy's Largest Root, along with their
respective F-statistics and p-values
}
\description{
Extracts the multivariate test results from an ANOVA object.
}
\examples{
model <- ANOVA(data = mydata, dv = "score", formula = ~ group)
multi(model)

}
