% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.lmer.r
\name{r2.lmer}
\alias{r2.lmer}
\title{Calculate R-squared based on correlation between fitted values and observed values}
\usage{
r2.lmer(m)
}
\description{
Calculate R-squared based on correlation between fitted values and observed values
}
