% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNA.R
\name{removeNA}
\alias{removeNA}
\title{Replace NA Values with Mean}
\usage{
removeNA(x)
}
\arguments{
\item{x}{A numeric vector containing NA values}
}
\value{
A numeric vector with NA values replaced by the mean
}
\description{
Replaces NA values in a numeric vector with the mean of non-NA values. 
Issues a warning if the proportion of NA values exceeds 5%.
}
\details{
The function:
* Calculates the proportion of NA values in the input vector
* Issues a warning if more than 5% of values are NA
* Replaces all NA values with the mean of non-NA values
}
\examples{
x <- c(1, 2, NA, 4, NA, 6)
removeNA(x)  # Returns c(1, 2, 3.25, 4, 3.25, 6)

}
