% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnormlim.r
\name{rnormlim}
\alias{rnormlim}
\title{Generate Random Normal Numbers with Limits}
\usage{
rnormlim(n, mu = 0, sigma = 1, min = -Inf, max = Inf)
}
\arguments{
\item{n}{Number of observations to generate}

\item{mu}{Mean of the normal distribution (default = 0)}

\item{sigma}{Standard deviation of the normal distribution (default = 1)}

\item{min}{Minimum allowed value (default = -Inf)}

\item{max}{Maximum allowed value (default = Inf)}
}
\value{
A numeric vector of length n containing random normal values 
constrained between min and max
}
\description{
Generates random numbers from a normal distribution with specified mean and 
standard deviation, constrained between minimum and maximum values. Any values 
outside the specified range are resampled until they fall within the limits.
}
\examples{
# Generate 1000 standard normal values between -2 and 2
x <- rnormlim(1000, mu = 0, sigma = 1, min = -2, max = 2)

# Generate 500 normal values with mean 10, sd 2, minimum 5
y <- rnormlim(500, mu = 10, sigma = 2, min = 5)

}
