% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{se}
\alias{se}
\title{Standard Error}
\usage{
se(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Logical; should NA values be removed?}
}
\value{
Standard error value
}
\description{
Calculates the standard error of a numeric vector.
Adapted from the sciplot package.
}
\examples{
se(c(1,2,3,4,5)) # Returns standard error
se(c(1,2,NA,4,5), na.rm=TRUE) # Handles NA values

}
