% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapelab.install.r
\name{shapelab.install}
\alias{shapelab.install}
\title{Install shapelab package}
\usage{
shapelab.install(address = "~/Dropbox/RtG Simulator/analysis r/shapelab/",
  shapelab.type = "source")
}
\description{
Install shapelab package
}
