% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.experiment.r
\name{simulate.experiment}
\alias{simulate.experiment}
\title{simulate an experiment}
\usage{
simulate.experiment(object, subj, repetitions, return_data = T)
}
\arguments{
\item{object}{an object of the class shapelab.exp.design produced with the function design.experiment}

\item{subj}{number of subjects}

\item{repetitions}{number of repetitions per subject}

\item{return_data}{default to TRUE. Use FALSE to produce just a trials sequence}
}
\value{
An object of the class shapelab.exp.design with the following values

seed the seed that was used during the simulation

exp the simulated data

coef.matrix the subj-by-subj coefficient matrix
}
\description{
simulate an experiment
}
\examples{
libraries()

# set a seed
trial.seed <- 9000

# initialize the factors list
lIV <- list(
  Var1 = factor(1:2, labels = c("a","b"))
  ,
  Var2 = factor(1:2)
)

tIV <- c("w","b")

# describe the model
exp.formula <- ~ Var1*Var2

# initialize the coefficients
cus.beta <- c(1, 4, 5, 1)*3 # means of the fixed effects
cus.theta <- c(1, 4, 5, 1) # standard errors of the random components
cus.sigma <- 3 # sigma of the model

# few examples
(r <- design.experiment(lIV, tIV, betas = cus.beta, thetas = cus.theta, s = cus.sigma, exp.formula, seed=trial.seed))

# start the simulation
exp.info <- simulate.experiment(r, 10, 10)

# extract the simulated data
exp <- exp.info$exp

# plot the results
ggplot(aes(Var1, y, color=Var2, group=Var2), data=exp) + facet_wrap(~subjName) + 
  geom_point(size=3, alpha=.4) +                                                              # draw raw scores
  stat_smooth(method='lm', level=.68, se=F, size=1) +                                         # draw fit line
  stat_summary(fun.data='mean_sdl', geom='point', size=3, color='black') +                    # draw mean points
  stat_summary(fun.data='mean_sdl', mult=1/sqrt(10), geom='errorbar', width=.2, size=1.2) +   # draw errorbars
  theme_bw()
  
# check a fit of the data (with more subjects)
exp.info <- simulate.experiment(r, 100, 10)
exp <- exp.info$exp
(lmod1 <- lmer(y ~ Var1*Var2 + (1 + Var1*Var2|subjName), data=exp))
}
