% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{sortdataframe}
\alias{sortdataframe}
\title{Sort Data Frame}
\usage{
sortdataframe(x, decreasing = FALSE, by = 1, ...)
}
\arguments{
\item{x}{A data frame}

\item{decreasing}{Logical; should the sort be in decreasing order?}

\item{by}{Column name(s) or number(s) to sort by}

\item{...}{Additional arguments passed to order()}
}
\value{
Sorted data frame
}
\description{
Sorts a data frame by one or more columns specified by name or number
}
\examples{
df <- data.frame(x=c(2,1,3), y=c("b","a","c"))
sortdataframe(df, by="x") # Sort by x column
sortdataframe(df, by=c("x","y")) # Sort by multiple columns

}
