% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subj.repetitions.find.r
\name{subj.repetition.find}
\alias{subj.repetition.find}
\title{Find the ideal amount of subjects and of repetitions per subject}
\usage{
subj.repetition.find(object, find.subj = T, find.reps = T, subj.start = 2,
  subj.max = 20, reps.start = 2, reps.max = 15)
}
\arguments{
\item{object}{an object of the type shapelab.exp.design as generated with design.experiment}

\item{find.subj}{logical: do you want to determine the ideal amount of subjects? (default to TRUE)}

\item{find.reps}{logical: do you want to determine the ideal amount of repetitions per subject? (default to TRUE)}

\item{subj.start}{initial amount of subjects for the simulation}

\item{subj.max}{maximum amount of subjects for the simulation}

\item{reps.start}{initial amount of repetitions per subject for the simulation}

\item{reps.max}{maximum amount of repetitions per subject for the simulation}
}
\description{
This function is a tool to help pre-determining the amount of subjects and/or repetitions per subject given the a-priori distribution of the parameters of a model
}
