% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialRepair.r
\name{trialRepair}
\alias{trialRepair}
\title{Repair the extraction of kinematics}
\usage{
trialRepair(n_trial, trajData, kinData, modifications = NULL,
  MGA.range = NULL, TGA.type = "s", TGA.lim = 20, frame = T, ...)
}
\arguments{
\item{n_trial}{number of the trial to repair}

\item{trajData}{dataset with trajectory data}

\item{kinData}{dataset with the extracted kinematics as produced by kin.extract}

\item{modifications}{in quotes, the modifications to apply in the new extraction of the kinematics (see example)}

\item{MGA.range}{as in kin.extract}

\item{TGA.type}{as in kin.extract}

\item{TGA.lim}{as in kin.extract}

\item{frame}{logic: should the data be plotted against frames (TRUE) or time (FALSE)}
}
\examples{

data(rtgData)
# always check the data first
rtgData <- data.check(rtgData)

# make sure that the trial count starts from 1 and not from zero
rtgData <- cc.trialN(rtgData)
# smooth and derive the dataset
rtgData <- kin.SmoothAndDerive(rtgData)
# extract the kinematics
kinData <- kin.extract(subset(rtgData, GPZ>100 & GA>15 & frameN > 10), graphs=F, TGA.type="s", TGA.lim = 20)
# view all trials at once
view.summary.trials.GA(datatraj = rtgData, datakin = kinData, print =F)

# repair trials
repaired_trial <- trialRepair(3, rtgData, kinData, "GPZ>100 & GA>15 & frameN > 10", MGA.range = c(70, 90))
# this displays the modifications and returns the relative line of kinData BUT does not modifies kinData
view.single.trial.GA(3, rtgData, kinData) # if you plot again the trial using kinData, you can see it is still unmodified
# in order to apply the modifications to kinData permanently, you need to rewrite that line in it
kinData[kinData$trialN==3,] <- repaired_trial
view.single.trial.GA(3, rtgData, kinData) # not it shows up correctly

# and so on
repaired_trial <- trialRepair(41, rtgData, kinData, "GPZ>100 & GA>15 & frameN > 10", MGA.range = c(10, 50), TGA.lim = 35)
kinData[kinData$trialN==41,] <- repaired_trial
# it is always preferable to first take a look at the preview without overwriting kinData immediately

# finally, it might be convenient to set a variable with the number of the trial
trial_to_repair <- 44
repaired_trial <- trialRepair(trial_to_repair, rtgData, kinData, "GPZ>100 & GA>15 & frameN > 10", MGA.range = c(70, 110), TGA.type = "v", TGA.lim = .5)
kinData[kinData$trialN==trial_to_repair,] <- repaired_trial

detach(rtgData)

}
