% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA_utils.R
\name{uni}
\alias{uni}
\title{Extract Univariate Results from ANOVA Object}
\usage{
uni(ANOVA.object)
}
\arguments{
\item{ANOVA.object}{An object of class "ANOVA" containing analysis results}
}
\value{
A data frame containing univariate test results including F-values, 
degrees of freedom, p-values, and effect sizes
}
\description{
Extracts the univariate test results from an ANOVA object.
}
\examples{
model <- ANOVA(data = mydata, dv = "score", formula = ~ group)
uni(model)

}
