% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.library.r
\name{mu.library}
\alias{mu.library}
\title{Load libraries and download them if not existing}
\usage{
mu.library(wants)
}
\arguments{
\item{wants}{a vector of strings with the names of the packages to load}
}
\description{
Load libraries and download them if not existing
}
