% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.library.r
\name{cc.library}
\alias{cc.library}
\title{Load libraries and download them if not existing}
\usage{
cc.library(wants)
}
\arguments{
\item{wants}{a vector of strings with the names of the packages to load}
}
\description{
Load libraries and download them if not existing
}
